requirements.txt
altgraph==0.17.4
asttokens==3.0.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
coverage==7.10.6
coverage-conditional-plugin==0.9.0
decorator==5.2.1
docutils==0.22.2
executing==2.2.1
filetype==1.2.0
greenlet==3.2.4
idna==3.10
ImageHash==4.3.2
iniconfig==2.1.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Kivy==2.3.1
Kivy-Garden==0.1.5
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
numpy==2.3.3
packaging==25.0
parso==0.8.5
pexpect==4.9.0
piexif==1.1.3
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
pyheif==0.8.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
pytest==8.4.2
PyWavelets==1.9.0
requests==2.32.5
rich==14.1.0
scipy==1.16.2
setuptools==80.9.0
shellingham==1.5.4
SQLAlchemy==2.0.43
stack-data==0.6.3
traitlets==5.14.3
typer==0.19.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
